requirements.txt
numpy==1.23.1
opencv_contrib_python==4.9.0.80
opencv_python==4.8.0.76
opencv_python_headless==4.8.0.76
Pillow==10.0.1
Pillow==11.0.0
scipy==1.14.1
torch==1.12.0+cu116
torchvision==0.13.0+cu116
